+++ functions/src/handlers/factory/handler-factory.ts
@@ -1,8 +1,9 @@
 import { APIGatewayEvent, Context } from 'aws-lambda';
 import { logger } from '../../common/logger';
-import { HttpStatus, LambdaResponse, httpResponse } from './http-response';
+import { LambdaResponse, httpResponse } from '../http/http-response';
 import { ErrorCode } from '../../common/error-codes';
-import { AppError, ClientError } from '../../common/app-errors';
+import { AppError } from '../../common/app-errors';
+import { httpErrorHandler } from './http-error-handler';
 
 export type RequestHandler = (
   event: APIGatewayEvent,
@@ -16,14 +17,14 @@ export type RequestErrorHandler = (error: AppError) => Promise<LambdaResponse>;
 export const handlerFactory = (
   name: string,
   requestHandler: RequestHandlerWithoutContext,
-  requestErrorHandler: RequestErrorHandler = defaultErrorHandler,
+  errorHandler: RequestErrorHandler = httpErrorHandler,
 ): RequestHandler => {
   return async (event, context) => {
     try {
       logger.addContext(context);
       return await requestHandlerWithLog(name, requestHandler, event);
     } catch (e: unknown) {
-      return await requestErrorHandlerWithLog(name, requestErrorHandler, e);
+      return await requestErrorHandlerWithLog(name, errorHandler, e);
     }
   };
 };
@@ -53,38 +54,5 @@ const requestErrorHandlerWithLog = async (
   } else {
     logger.error(`An unexpected error occurred in handler: ${name}`, String(e));
     return httpResponse(500).withError(ErrorCode.UNKNOWN_ERROR);
-  }
-};
-
-const defaultErrorHandler: RequestErrorHandler = async (
-  error: AppError,
-): Promise<LambdaResponse> => {
-  switch (error.code) {
-    case ErrorCode.TASK_NOT_FOUND:
-      return httpResponse(HttpStatus.NOT_FOUND).withError(
-        ErrorCode.TASK_NOT_FOUND,
-      );
-    case ErrorCode.DDB_CLIENT_ERROR:
-      return httpResponse(HttpStatus.BAD_REQUEST).withError(
-        ErrorCode.DDB_CLIENT_ERROR,
-      );
-    case ErrorCode.DDB_SERVER_ERROR:
-      return httpResponse(HttpStatus.INTERNAL_SERVER_ERROR).withError(
-        ErrorCode.DDB_SERVER_ERROR,
-      );
-    case ErrorCode.DDB_UNKNOWN_ERROR:
-      return httpResponse(HttpStatus.INTERNAL_SERVER_ERROR).withError(
-        ErrorCode.DDB_UNKNOWN_ERROR,
-      );
-    default:
-      if (error instanceof ClientError) {
-        return httpResponse(HttpStatus.BAD_REQUEST).withError(
-          ErrorCode.INVALID_REQUEST,
-        );
-      } else {
-        return httpResponse(HttpStatus.INTERNAL_SERVER_ERROR).withError(
-          ErrorCode.UNKNOWN_ERROR,
-        );
-      }
   }
 };
