+++ functions/src/usecases/create-task-usecase.ts
@@ -4,7 +4,7 @@ import {
   createTask as ddbCreateTask,
 } from '../infrastructure/ddb/tasks-table';
 import { TaskUnknownError } from '../domain/errors/task-errors';
-import { useCaseFactory } from './utils/usecase-factory';
+import { useCaseFactory } from './factory/usecase-factory';
 import { CreateTaskRequest } from '../handlers/request_schemas/create-task-request';
 
 const createTask = async (body: CreateTaskRequest): Promise<Task> => {
